+++ MainProject/SingleBezierSurfaceC0.cpp
@@ -93,7 +93,7 @@ void SingleBezierSurfaceC0::Accept(AbstractModelVisitor& visitor)
 	visitor.Accept(IModel::downcasted_shared_from_this<SingleBezierSurfaceC0>());
 }
 
-DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetValue(float u, float v)
+DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetValue(double u, double v)
 {
 	std::array<Vector3, 4> vertical;
 	for (int i = 0; i < 4; i++)
@@ -110,7 +110,7 @@ DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetValue(float u, float v)
 	return result;
 }
 
-DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetUDerivativeValue(float u, float v)
+DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetUDerivativeValue(double u, double v)
 {
 	std::array<Vector3, 4> horizontal;
 	for (int j = 0; j < 4; j++)
@@ -127,7 +127,7 @@ DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetUDerivativeValue(float u,
 	return result;
 }
 
-DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetVDerivativeValue(float u, float v)
+DirectX::SimpleMath::Vector3 SingleBezierSurfaceC0::GetVDerivativeValue(double u, double v)
 {
 	std::array<Vector3, 4> vertical;
 	for (int i = 0; i < 4; i++)
