+++ MainProject/SingleBezierSurfaceC2.h
@@ -21,13 +21,13 @@ public:
 
 
 	// Inherited via ISingleBezierSurface
-	virtual DirectX::SimpleMath::Vector3 GetValue(float u, float v) override;
+	virtual DirectX::SimpleMath::Vector3 GetValue(double u, double v) override;
 
 
-	virtual DirectX::SimpleMath::Vector3 GetUDerivativeValue(float u, float v) override;
+	virtual DirectX::SimpleMath::Vector3 GetUDerivativeValue(double u, double v) override;
 
 
-	virtual DirectX::SimpleMath::Vector3 GetVDerivativeValue(float u, float v) override;
+	virtual DirectX::SimpleMath::Vector3 GetVDerivativeValue(double u, double v) override;
 
 
 	virtual bool IsUWrapped() override;
